Split Error in a module
